+++ core/converters/converters.go
@@ -1,4 +1,4 @@
-package converter
+package converters
 
 import (
 	"encoding/hex"
